petstore/requirements.txt
pytest==7.1.2
requests==2.28.1
allure-pytest==2.9.45
pydantic==1.9.2
faker==14.1.0